+++ src/common/js/i18n.js
@@ -2,6 +2,8 @@
  * pwix:modal/src/common/js/i18n.js
  */
 
+import { pwixI18n } from 'meteor/pwix:i18n';
+
 pwixModal.i18n = {
     en_US: {
         MD_BUTTON_OK: 'OK',
@@ -24,3 +26,12 @@ pwixModal.i18n = {
 // actually a package should only provide these short language fallbacks
 pwixModal.i18n.en = pwixModal.i18n.en_US;
 pwixModal.i18n.fr = pwixModal.i18n.fr_FR;
+
+pwixI18n.namespace( I18N, pwixModal.i18n );
+
+/**
+ * @returns {String} the i18n namespace of the package
+ */
+pwixModal.i18n.namespace = function(){
+    return I18N;
+};